+++ sdk/storage/azure-storage-blobs/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Added support for telemetry options.
 - Added `Azure::Storage::Blobs::PackageVersion`.
+- Added `ShareFileClient::GetRangeListDiff`.
 
 ### Breaking Changes
 
@@ -20,6 +21,7 @@
 - Changed return type of `BlobContainerClient::GetAccessPolicy` to `BlobContainerAccessPolicy`.
 - Changed return type of `BlobClient::GetProperties` to `BlobProperties`.
 - Changed return type of `BlobClinet::GetTags` to `std::map<std::string, std::string>`.
+- Removed `PreviousShareSnapshot` from `GetShareFileRangeListOptions`, use `ShareFileClient::GetRangeListDiff` instead.
 
 ### Other Changes and Improvements